Merge/v2 master (#44)
* fix reactive objects

* Fix an enum previously renamed and .value in ref.

* Add front update events

* Fix merge conflicts

* Fix merge conflicts

* ...

* Fix: merge conflicts.

* feat: adminWatch, userWatch

* Fix: setCurrentChannel -> setUserCurrentChannel

* Fix: Update User.name to point to User.nickname

* Revert last commit

Fix: Update User.name to point to User.nickname

* Fix: EventUser to aggregate user information for an event

* fix: multiple broken stuff after merge with chatv2

* style

* wip: still not working but at least it doesn't crash

* feat: create relation block user

* feat: new channel_ban table. add block relations

* feat: add new Ban entity with User and Channel relations

* merge with chatv4

* feat: join/part/kick now update client

* Fix merge conflicts

* fix: user section null when not selected channel

* feat: fix right click on users (in event and users section)

* feat: add challenge and spectate buttons. fix: issue when first loading the chat channel not selected

* feat: challenge system

* fix: game service change user properties because of properties name changes

* wip: spectate

* fix: wild dot

* fix: ups

* fix: Removed duplicated function

* return .env to localhost

* feat: Set TRANSC_HOST value from Makefile configuration

* fix: TRANSC_HOST

* Fix: spectate event name

* fix: typo

* Fix: frontend spectate rename to challengeSpectate

* fix: spectate event working on game backend

* chore: let game handle ingame user status

* feat: spectate working

* fix: chat-client not getting notified of challengespectated

* fix: frontend ChannelUser properties reactive

* feat: add promote/demote and ban/unban options

* chore: remove unneccesary log

* fix: Resolved issue with Mac understanding configuration settings

* feat: challenge working

* Fix: userWatch frontend functionallity

* feat: add onDelete cascade in block entity

* fix: change delete for remove in block remove

* Fixing getting user status

* fix: no spaces on target chat events

* fix: add friend and blocked to user roles

* feat: send game backend notice we are on menu for status update

* Fix: now handles exception when empty cookies

* fix: createBan and removeBan working

* feat:
  - findAllWithBlocksAndFriends in User
  - findAllWithUsersAndBans in Channel

* fix: user accepting challenge while in queue

* Fix: merge conflicts

* Fix: watch error

* feat: add modal

* Add intra requests limiter (#37)

* Fix: Merge conflicts

* chore: add message for when no channel are available

* fix: password protected channels now working

* Fix: SummaryChannelDTO.password value

* fix: add encrypt password in channel creation

* Fix: channel password list

* fix: blank character password when joining channel

* fixes

* Checking status

* Status done

* fix: add events in next tick as a measure to mitigate a not recreable error

* fix: message input not visible if not in a channel

* Perfil funcional

* feat: base friendsview component (not working yet)

---------

+++ src/requirements/backend/vol/src/channels/channels.service.ts
@@ -14,6 +14,7 @@ import { CreateChannelUserDto } from './dto/create-channel-user.dto';
 import { ChannelUserModeDto } from './dto/channel-user-mode.dto';
 
 import { User } from '../users/entities/user.entity';
+import { Ban } from '../users/entities/ban.entity';
 import * as bcrypt from 'bcrypt';
 
 @Injectable()
@@ -32,7 +33,16 @@ export class ChannelsService {
   ) {}
 
   async create(createChannelDto: CreateChannelDto): Promise<Channel> {
-    const newChannel = new Channel(createChannelDto);
+    if (createChannelDto.password != undefined) {
+      createChannelDto.password = await this.encryptPassword(createChannelDto.password);
+    }
+    const newChannel = new Channel(
+      createChannelDto.name,
+      createChannelDto.ownerId,
+      createChannelDto.id,
+      createChannelDto.topic,
+      createChannelDto.password
+    );
     await this.channelsRepository.save(newChannel);
 
     newChannel.users = [
@@ -42,6 +52,7 @@ export class ChannelsService {
         admin: false
       })
     ];
+    newChannel.bans = [];
     return newChannel;
   }
 
@@ -59,14 +70,34 @@ export class ChannelsService {
 
   async	findAllWithUsers(): Promise<Channel[]> {
     return await this.channelsRepository.find({
-      relations: ['users', 'users.channel', 'users.user'],
+      relations: ['users', 'users.user'],
+    })
+  }
+
+  async	findAllWithUsersAndBans(): Promise<Channel[]> {
+    return await this.channelsRepository.find({
+      relations: ['users', 'users.user', 'bans.user'],
     })
   }
 
   async remove(id: string) {
     await this.channelsRepository.delete(id);
   }
+/*
+  async getBansByChannel(channelId: string) {
+    const channel = await this.channelsRepository.findOne({
+      relations: ['bans'],
+      where: {
+        id: channelId,
+      },
+    });
+    if (!channel) {
+      throw new HttpException('Channel not found', HttpStatus.NOT_FOUND);
+    }
 
+    return channel.bans;
+  }
+*/
   async createChannelUser(createChannelUserDto: CreateChannelUserDto): Promise<ChannelUser> {
     const channel = await this.channelsRepository.findOne({
       relations: ['users'],
@@ -85,6 +116,7 @@ export class ChannelsService {
       }
     });
     if (!user) {
+      console.log(createChannelUserDto);
       throw new HttpException('User not found', HttpStatus.NOT_FOUND);
     }
 
@@ -105,7 +137,16 @@ export class ChannelsService {
   }
 
   async removeChannelUser(channelId: string, userId: string): Promise<Channel> {
-    const channel = await this.findOneById(channelId);
+    const channel = await this.channelsRepository.findOne({
+      relations: ['users'],
+      where: {
+        id: channelId,
+      }
+    });
+    if (!channel) {
+      throw new HttpException('Channel not found', HttpStatus.NOT_FOUND);
+    }
+
     const channelUser = await this.findChannelUser(channelId, userId);
 
     await this.channelUsersRepository.delete(channelUser);
@@ -169,12 +210,16 @@ export class ChannelsService {
     return this.channelsRepository.save(channel);
   }
 
+  async encryptPassword(password: string) : Promise<string> {
+    const salt = await bcrypt.genSalt(10);
+    const hashedPassword = await bcrypt.hash(password, salt);
+    return hashedPassword;
+  }
+
   async setChannelPassword(channelId: string, password: string): Promise<Channel> {
     const channel = await this.findOneById(channelId);
 
-    const salt = await bcrypt.genSalt(10);
-    const hashedPassword = await bcrypt.hash(password, salt);
-    channel.password = hashedPassword;
+    channel.password = await this.encryptPassword(password);
     return this.channelsRepository.save(channel);
   }
 
@@ -201,16 +246,6 @@ export class ChannelsService {
     );
 
     channelUser.admin = channelUserModeDto.mode;
-    return this.channelUsersRepository.save(channelUser);
-  }
-
-  async setBannedToChannelUser(channelUserModeDto: ChannelUserModeDto): Promise<ChannelUser> {
-    const channelUser = await this.findChannelUser(
-      channelUserModeDto.channelId,
-      channelUserModeDto.userId
-    );
-
-    channelUser.banned = channelUserModeDto.mode;
     return this.channelUsersRepository.save(channelUser);
   }
